Moved function protocols to interfaces
