+++ frictionless/check.py
@@ -7,6 +7,7 @@ if TYPE_CHECKING:
     from .row import Row
     from .error import Error
     from .resource import Resource
+    from .interfaces import CheckFunction
 
 
 class Check(Metadata):
@@ -82,12 +83,3 @@ class Check(Metadata):
     # Metadata
 
     metadata_Error = errors.CheckError
-
-
-# Internal
-
-
-# TODO: add to interfaces?
-class CheckFunction(Protocol):
-    def __call__(self, row: Row) -> Iterable[Error]:
-        ...
